requirements.txt
streamlit
crewai
python-docx
PyMuPDF
